+++ src/algorithms/npag.rs
@@ -12,9 +12,11 @@ use crate::structs::theta::Theta;
 use anyhow::bail;
 use anyhow::Result;
 use pharmsol::prelude::{
-    data::{Data, ErrorModel},
+    data::{Data, ErrorModels},
     simulator::Equation,
 };
+
+use pharmsol::prelude::ErrorModel;
 
 use faer::Col;
 
@@ -41,8 +43,8 @@ pub struct NPAG<E: Equation> {
     f0: f64,
     f1: f64,
     cycle: usize,
-    gamma_delta: f64,
-    error_model: ErrorModel,
+    gamma_delta: Vec<f64>,
+    error_models: ErrorModels,
     converged: bool,
     status: Status,
     cycle_log: CycleLog,
@@ -65,8 +67,8 @@ impl<E: Equation> Algorithms<E> for NPAG<E> {
             f0: -1e30,
             f1: f64::default(),
             cycle: 0,
-            gamma_delta: 0.1,
-            error_model: settings.error().clone().into(),
+            gamma_delta: vec![0.1; settings.errormodels().len()],
+            error_models: settings.errormodels().clone().into(),
             converged: false,
             status: Status::Starting,
             cycle_log: CycleLog::new(),
@@ -169,7 +171,7 @@ impl<E: Equation> Algorithms<E> for NPAG<E> {
             delta_objf: (self.last_objf - self.objf).abs(),
             nspp: self.theta.nspp(),
             theta: self.theta.clone(),
-            gamlam: self.error_model.scalar(),
+            error_models: self.error_models.clone(),
             status: self.status.clone(),
         };
 
@@ -187,7 +189,7 @@ impl<E: Equation> Algorithms<E> for NPAG<E> {
             &self.equation,
             &self.data,
             &self.theta,
-            &self.error_model,
+            &self.error_models,
             self.cycle == 1 && self.settings.config().progress,
             self.cycle != 1,
         )?;
@@ -273,75 +275,97 @@ impl<E: Equation> Algorithms<E> for NPAG<E> {
     }
 
     fn optimizations(&mut self) -> Result<()> {
-        // Gam/Lam optimization
-        // TODO: Move this to e.g. /evaluation/error.rs
-        let gamma_up = self.error_model.scalar() * (1.0 + self.gamma_delta);
-        let gamma_down = self.error_model.scalar() / (1.0 + self.gamma_delta);
+        self.error_models
+            .clone()
+            .iter_mut()
+            .filter_map(|(outeq, em)| match em {
+                ErrorModel::None => None,
+                _ => Some((outeq, em)),
+            })
+            .try_for_each(|(outeq, em)| -> Result<()> {
+                // OPTIMIZATION
 
-        let mut error_model_up = self.error_model.clone();
-        error_model_up.set_scalar(gamma_up);
+                let gamma_up = em.scalar()? * (1.0 + self.gamma_delta[outeq]);
+                let gamma_down = em.scalar()? / (1.0 + self.gamma_delta[outeq]);
 
-        let mut error_model_down = self.error_model.clone();
-        error_model_down.set_scalar(gamma_down);
+                let mut error_model_up = self.error_models.clone();
+                error_model_up.set_scalar(outeq, gamma_up)?;
 
-        let psi_up = calculate_psi(
-            &self.equation,
-            &self.data,
-            &self.theta,
-            &error_model_up,
-            false,
-            true,
-        )?;
-        let psi_down = calculate_psi(
-            &self.equation,
-            &self.data,
-            &self.theta,
-            &error_model_down,
-            false,
-            true,
-        )?;
+                let mut error_model_down = self.error_models.clone();
+                error_model_down.set_scalar(outeq, gamma_down)?;
 
-        let (lambda_up, objf_up) = match burke(&psi_up) {
-            Ok((lambda, objf)) => (lambda, objf),
-            Err(err) => {
-                //todo: write out report
-                return Err(anyhow::anyhow!("Error in IPM during optim: {:?}", err));
-            }
-        };
-        let (lambda_down, objf_down) = match burke(&psi_down) {
-            Ok((lambda, objf)) => (lambda, objf),
-            Err(err) => {
-                //todo: write out report
-                //panic!("Error in IPM: {:?}", err);
-                return Err(anyhow::anyhow!("Error in IPM during optim: {:?}", err));
-                //(Array1::zeros(1), f64::NEG_INFINITY)
-            }
-        };
-        if objf_up > self.objf {
-            self.error_model.set_scalar(gamma_up);
-            self.objf = objf_up;
-            self.gamma_delta *= 4.;
-            self.lambda = lambda_up;
-            self.psi = psi_up;
-        }
-        if objf_down > self.objf {
-            self.error_model.set_scalar(gamma_down);
-            self.objf = objf_down;
-            self.gamma_delta *= 4.;
-            self.lambda = lambda_down;
-            self.psi = psi_down;
-        }
-        self.gamma_delta *= 0.5;
-        if self.gamma_delta <= 0.01 {
-            self.gamma_delta = 0.1;
-        }
+                let psi_up = calculate_psi(
+                    &self.equation,
+                    &self.data,
+                    &self.theta,
+                    &error_model_up,
+                    false,
+                    true,
+                )?;
+                let psi_down = calculate_psi(
+                    &self.equation,
+                    &self.data,
+                    &self.theta,
+                    &error_model_down,
+                    false,
+                    true,
+                )?;
+
+                let (lambda_up, objf_up) = match burke(&psi_up) {
+                    Ok((lambda, objf)) => (lambda, objf),
+                    Err(err) => {
+                        //todo: write out report
+                        return Err(anyhow::anyhow!("Error in IPM during optim: {:?}", err));
+                    }
+                };
+                let (lambda_down, objf_down) = match burke(&psi_down) {
+                    Ok((lambda, objf)) => (lambda, objf),
+                    Err(err) => {
+                        //todo: write out report
+                        //panic!("Error in IPM: {:?}", err);
+                        return Err(anyhow::anyhow!("Error in IPM during optim: {:?}", err));
+                        //(Array1::zeros(1), f64::NEG_INFINITY)
+                    }
+                };
+                if objf_up > self.objf {
+                    self.error_models.set_scalar(outeq, gamma_up)?;
+                    self.objf = objf_up;
+                    self.gamma_delta[outeq] *= 4.;
+                    self.lambda = lambda_up;
+                    self.psi = psi_up;
+                }
+                if objf_down > self.objf {
+                    self.error_models.set_scalar(outeq, gamma_down)?;
+                    self.objf = objf_down;
+                    self.gamma_delta[outeq] *= 4.;
+                    self.lambda = lambda_down;
+                    self.psi = psi_down;
+                }
+                self.gamma_delta[outeq] *= 0.5;
+                if self.gamma_delta[outeq] <= 0.01 {
+                    self.gamma_delta[outeq] = 0.1;
+                }
+                Ok(())
+            })?;
+
         Ok(())
     }
 
     fn logs(&self) {
         tracing::info!("Objective function = {:.4}", -2.0 * self.objf);
         tracing::debug!("Support points: {}", self.theta.nspp());
-        tracing::debug!("Gamma = {:.16}", self.error_model.scalar());
+
+        self.error_models.iter().for_each(|(outeq, em)| {
+            if ErrorModel::None == *em {
+                return;
+            }
+            tracing::debug!(
+                "Error model for outeq {}: {:.16}",
+                outeq,
+                em.scalar().unwrap_or_default()
+            );
+        });
+
         tracing::debug!("EPS = {:.4}", self.eps);
         // Increasing objf signals instability or model misspecification.
         if self.last_objf > self.objf + 1e-4 {