+++ src/routines/output.rs
@@ -553,7 +553,7 @@ impl<E: Equation> NPResult<E> {
 pub struct NPCycle {
     pub cycle: usize,
     pub objf: f64,
-    pub gamlam: f64,
+    pub error_models: ErrorModels,
     pub theta: Theta,
     pub nspp: usize,
     pub delta_objf: f64,
@@ -564,7 +564,7 @@ impl NPCycle {
     pub fn new(
         cycle: usize,
         objf: f64,
-        gamlam: f64,
+        error_models: ErrorModels,
         theta: Theta,
         nspp: usize,
         delta_objf: f64,
@@ -573,7 +573,7 @@ impl NPCycle {
         Self {
             cycle,
             objf,
-            gamlam,
+            error_models,
             theta,
             nspp,
             delta_objf,
@@ -585,7 +585,7 @@ impl NPCycle {
         Self {
             cycle: 0,
             objf: 0.0,
-            gamlam: 0.0,
+            error_models: ErrorModels::default(),
             theta: Theta::new(),
             nspp: 0,
             delta_objf: 0.0,
@@ -621,8 +621,23 @@ impl CycleLog {
         writer.write_field("converged")?;
         writer.write_field("status")?;
         writer.write_field("neg2ll")?;
-        writer.write_field("gamlam")?;
         writer.write_field("nspp")?;
+        if let Some(first_cycle) = self.cycles.first() {
+            first_cycle.error_models.iter().try_for_each(
+                |(outeq, errmod): (usize, &ErrorModel)| -> Result<(), csv::Error> {
+                    match errmod {
+                        ErrorModel::Additive { .. } => {
+                            writer.write_field(format!("gamlam.{}", outeq))?;
+                        }
+                        ErrorModel::Proportional { .. } => {
+                            writer.write_field(format!("gamlam.{}", outeq))?;
+                        }
+                        ErrorModel::None { .. } => {}
+                    }
+                    Ok(())
+                },
+            )?;
+        }
 
         let parameter_names = settings.parameters().names();
         for param_name in &parameter_names {
@@ -638,10 +653,25 @@ impl CycleLog {
             writer.write_field(format!("{}", cycle.status == Status::Converged))?;
             writer.write_field(format!("{}", cycle.status))?;
             writer.write_field(format!("{}", cycle.objf))?;
-            writer.write_field(format!("{}", cycle.gamlam))?;
             writer
-                .write_field(format!("{}", cycle.theta.matrix().nrows()))
+                .write_field(format!("{}", cycle.theta.nspp()))
                 .unwrap();
+
+            // Write the error models
+            cycle.error_models.iter().try_for_each(
+                |(_, errmod): (usize, &ErrorModel)| -> Result<()> {
+                    match errmod {
+                        ErrorModel::Additive { .. } => {
+                            writer.write_field(format!("{:.5}", errmod.scalar()?))?;
+                        }
+                        ErrorModel::Proportional { .. } => {
+                            writer.write_field(format!("{:.5}", errmod.scalar()?))?;
+                        }
+                        ErrorModel::None { .. } => {}
+                    }
+                    Ok(())
+                },
+            )?;
 
             for param in cycle.theta.matrix().col_iter() {
                 let param_values: Vec<f64> = param.iter().cloned().collect();